+++ BitPayLib/Exception/class-bitpayinvalidorder.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace BitPayLib\Exception;
+
+/**
+ * Plugin Name: BitPay Checkout for WooCommerce
+ * Plugin URI: https://www.bitpay.com
+ * Description: BitPay Checkout Plugin
+ * Version: 6.0.0
+ * Author: BitPay
+ * Author URI: mailto:integrations@bitpay.com?subject=BitPay Checkout for WooCommerce
+ */
+class BitPayInvalidOrder extends \RuntimeException {
+}