+++ BitPayLib/class-bitpaycancelorder.php
@@ -56,6 +56,6 @@ class BitPayCancelOrder {
 	}
 
 	private function clear_cookie_for_invoice_id(): void {
-		setcookie( 'bitpay-invoice-id', '', time() - 3600 );
+		setcookie( BitPayPluginSetup::COOKIE_INVOICE_ID_NAME, '', time() - 3600 );
 	}
 }