+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 # 6.0.0
 * Fixed a bug when the IPN/webhook is received with "complete" status
+* Add Unit & End2End tests
 
 # 5.4.0
 * Added compatibility with Checkout Blocks